+++ README.md
@@ -5,7 +5,7 @@ Software package and ROS wrappers of the [Aruco][1] Augmented Reality marker det
 
 
 ### Features
-<img align="right" src="https://raw.github.com/pal-robotics/aruco_ros/master/etc/marker_in_hand.jpg" />
+<img align="right" src="https://raw.github.com/pal-robotics/aruco_ros/master/aruco_ros/etc/marker_in_hand.jpg" />
  * High-framerate tracking of AR markers
  * Generate AR markers with given size and optimized for minimal perceptive ambiguity (when there are more markers to track)
  * Enhanced precision tracking by using boards of markers
@@ -17,7 +17,7 @@ Software package and ROS wrappers of the [Aruco][1] Augmented Reality marker det
  * Object pose estimation
  * Visual servoing: track object and hand at the same time
 
-<img align="right" src="https://raw.github.com/pal-robotics/aruco_ros/master/etc/reem_gazebo_floating_marker_world.png"/>
+<img align="right" src="https://raw.github.com/pal-robotics/aruco_ros/master/aruco_ros/etc/reem_gazebo_floating_marker_world.png"/>
 
 ### ROS API
 
@@ -66,7 +66,7 @@ Software package and ROS wrappers of the [Aruco][1] Augmented Reality marker det
     rosrun image_view image_view image:=/aruco_single/result
     ```
 
-<img align="right" src="https://raw.github.com/pal-robotics/aruco_ros/master/etc/reem_gazebo_floating_marker.png"/>
+<img align="right" src="https://raw.github.com/pal-robotics/aruco_ros/master/aruco_ros/etc/reem_gazebo_floating_marker.png"/>
 
 
 [1]: http://www.sciencedirect.com/science/article/pii/S0031320314000235 "Automatic generation and detection of highly reliable fiducial markers under occlusion by S. Garrido-Jurado and R. Muñoz-Salinas and F.J. Madrid-Cuevas and M.J. Marín-Jiménez 2014"
