+++ aruco/cfg/ArucoThreshold.cfg
