+++ aruco/etc/marker26_5cm.jpg
