+++ aruco/etc/marker26_5cm_margin_1cm.jpg
