+++ aruco/etc/marker26_5cm_margin_2cm.jpg
