+++ aruco/etc/marker582_5cm.jpg
