+++ aruco/etc/marker582_5cm_margin_2cm.jpg
