+++ aruco/etc/marker_in_hand.jpg
