+++ aruco/etc/marker_in_object.jpg
