+++ aruco/etc/reem_gazebo_floating_marker.png
