+++ aruco/etc/reem_gazebo_floating_marker_world.png
