+++ aruco/include/aruco/aruco.h
@@ -128,7 +128,7 @@ For testing the applications, the library provides videos and the corresponding 
  
 */
 
-#include "markerdetector.h"
-#include "boarddetector.h"
-#include "cvdrawingutils.h"
+#include <aruco/markerdetector.h>
+#include <aruco/boarddetector.h>
+#include <aruco/cvdrawingutils.h>
