+++ aruco/include/aruco/boarddetector.h
@@ -28,10 +28,10 @@ or implied, of Rafael Muñoz Salinas.
 #ifndef _Aruco_BoardDetector_H
 #define _Aruco_BoardDetector_H
 #include <opencv2/opencv.hpp>
-#include "exports.h"
-#include "board.h"
-#include "cameraparameters.h"
-#include "markerdetector.h"
+#include <aruco/exports.h>
+#include <aruco/board.h>
+#include <aruco/cameraparameters.h>
+#include <aruco/markerdetector.h>
 using namespace std;
 
 namespace aruco
