+++ aruco/include/aruco/cvdrawingutils.h
@@ -27,8 +27,8 @@ or implied, of Rafael Muñoz Salinas.
 ********************************/
 #ifndef _ArUco_DrawUtils_H_
 #define _ArUco_DrawUtils_H_
-#include "exports.h"
-#include "aruco.h"
+#include <aruco/exports.h>
+#include <aruco/aruco.h>
 namespace aruco
 {
   /**\brief A set of functions to draw in opencv images
