+++ aruco/launch/double.launch
