+++ aruco/src/aruco/arucofidmarkers.cpp
