+++ aruco/src/aruco/cameraparameters.cpp
