+++ aruco/src/aruco/cvdrawingutils.cpp
