+++ aruco/src/aruco/marker.cpp
