+++ aruco/src/aruco/markerdetector.cpp
