+++ aruco_msgs/msg/MarkerArray.msg
@@ -1,0 +1,2 @@
+Header header
+aruco_msgs/Marker[] markers
