+++ aruco_ros/cfg/ArucoThreshold.cfg
@@ -1,0 +1,14 @@
+#!/usr/bin/env python
+PACKAGE = "aruco_ros"
+
+from dynamic_reconfigure.parameter_generator_catkin import *
+
+gen = ParameterGenerator()
+
+gen.add("param1", int_t, 0, "blockSize parameter of cv::adaptiveThreshold", 4, 1, 15)
+gen.add("param2", int_t, 0, "C parameter of cv::adaptiveThreshold", 8, 1, 15)
+gen.add("normalizeImage", bool_t, 0, "normalizeImage", True)
+gen.add("dctComponentsToRemove", int_t, 0, "DCT components to remove", 2, 1, 4)
+gen.add("degree", int_t, 0, "Degree to rotate", 0, 0, 360)
+
+exit(gen.generate(PACKAGE, "double", "ArucoThreshold"))
