+++ aruco_ros/include/aruco_ros/aruco_ros_utils.h
@@ -1,7 +1,7 @@
 #ifndef ARUCO_ROS_UTILS_H
 #define ARUCO_ROS_UTILS_H
 
-#include "aruco/aruco.h"
+#include <aruco/aruco.h>
 #include <sensor_msgs/CameraInfo.h>
 #include <tf/transform_datatypes.h>
