+++ aruco_ros/launch/double.launch
@@ -1,0 +1,28 @@
+<launch>
+
+    <arg name="marker1Id"         default="582"/>
+    <arg name="marker2Id"         default="26"/>
+    <arg name="markerSize"        default="0.04"/>    <!-- in m -->
+    <arg name="eye"               default="left"/>
+    <arg name="dct_normalization" default="True" />
+    <arg name="dct_filter_size"   default="2" />
+    <arg name="marker1_frame"     default="marker_hand_frame" />
+    <arg name="marker2_frame"     default="marker_object_frame" />
+
+
+    <node pkg="aruco_ros" type="double" name="aruco_simple">    
+        <remap from="/camera_info" to="/stereo/$(arg eye)/camera_info" />
+        <remap from="/image" to="/stereo/$(arg eye)/image_rect_color" />
+        <param name="image_is_rectified" value="True"/>
+        <param name="marker_size" value="$(arg markerSize)"/>
+        <param name="marker_id1" value="$(arg marker1Id)"/>
+        <param name="marker_id2" value="$(arg marker2Id)"/>
+        <param name="normalizeImage" value="$(arg dct_normalization)"/>
+        <param name="dct_components_to_remove" value="$(arg dct_filter_size)"/>
+        <param name="parent_name" value="stereo_gazebo_$(arg eye)_camera_optical_frame"/>
+        <param name="child_name1" value="$(arg marker1_frame)" />
+        <param name="child_name2" value="$(arg marker2_frame)" />
+    </node>
+
+</launch>
+
