+++ aruco_ros/src/simple_double.cpp
