+++ aruco_ros/src/simple_single.cpp
