--- msg/MarkerArray.msg
@@ -1,2 +1,0 @@
-Header header
-aruco_ros/Marker[] markers
