--- package.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <name>aruco_ros</name>
-  <version>1.0.0</version>
-  <description>
-    The ARUCO Library has been developed by the Ava group of the Univeristy of Cordoba(Spain).
-    It provides real-time marker based 3D pose estimation using AR markers.
-  </description>
-  <maintainer email="bence.magyar@pal-robotics.com">Bence Magyar</maintainer>
-  <license>BSD</license>
-  <url type="website">http://www.uco.es/investiga/grupos/ava/node/26</url>
-  <author email="rmsalinas@uco.es">Rafael Muñoz Salinas</author>
-  <author email="bence.magyar@pal-robotics.com">Bence Magyar</author>
-
-  <buildtool_depend>catkin</buildtool_depend>
-  <build_depend>message_generation</build_depend>
-  <build_depend>cv_bridge</build_depend>
-  <build_depend>dynamic_reconfigure</build_depend>
-  <build_depend>geometry_msgs</build_depend>
-  <build_depend>image_transport</build_depend>
-  <build_depend>pal_vision_segmentation</build_depend>
-  <build_depend>roscpp</build_depend>
-  <build_depend>tf</build_depend>
-  <run_depend>message_runtime</run_depend>
-  <run_depend>cv_bridge</run_depend>
-  <run_depend>dynamic_reconfigure</run_depend>
-  <run_depend>geometry_msgs</run_depend>
-  <run_depend>image_transport</run_depend>
-  <run_depend>pal_vision_segmentation</run_depend>
-  <run_depend>roscpp</run_depend>
-  <run_depend>tf</run_depend>
-</package>
